Oldes_files/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.2
python-dotenv==1.1.1
email-validator==2.2.0
